Add end-to-end PostgreSQL connectivity tests to molecule
Extend molecule tests to verify PGBouncer can communicate with PostgreSQL:

- Add prepare.yml to install and configure PostgreSQL before PGBouncer
  - Supports Ubuntu (22.04, 24.04) and Rocky Linux 9
  - Creates test database, user, and configures pg_hba.conf
  - Sets password_encryption=md5 for PGBouncer compatibility

- Update converge.yml with test user configuration for PGBouncer

- Add new testinfra tests:
  - test_postgresql_direct_connectivity: verifies direct PostgreSQL connection
  - test_pgbouncer_postgresql_connectivity: verifies connection through PGBouncer

- Fix typo in test file path (etc/ -> /etc/)
- Rename test file from test_defatuls.py to test_defaults.py
