+++ molecule/default/converge.yml
@@ -3,6 +3,11 @@
   hosts: all
   become: true
 
+  vars:
+    test_db_user: testuser
+    test_db_password: testpassword
+    test_db_name: testdb
+
   pre_tasks:
     - name: Install prerequisites
       block:
@@ -15,5 +20,13 @@
 
   roles:
     - role: ansible-role-pgbouncer
+      vars:
+        pgbouncer_users:
+          - name: "{{ test_db_user }}"
+            pass: "{{ test_db_password }}"
+        pgbouncer_databases:
+          - name: "{{ test_db_name }}"
+            host: 127.0.0.1
+            port: 5432
       tags:
         - pgbouncer
